Rename Chat.parseText -> Chat.formatText
(And move it to its own file, chat-formatter.js)
